+++ src/modules/script-editor/kernel/script-editor-workflow-controller.ts
@@ -6,8 +6,7 @@ import { loadScriptEditorProjectFromFiles } from "../application/editor-project-
 import { serializeScriptEditorProjectToFiles } from "../application/editor-project-save";
 import { markScriptEditorProjectCompleteForExport } from "../application/project-completion-state";
 import { exportScriptEditorProjectToScenarioPackFiles } from "../application/runtime-pack-export";
-import { loadScriptEditorProjectFromScenarioPackUrl } from "../application/runtime-pack-import";
-import { DEFAULT_SCRIPT_EDITOR_TEMPLATE_SCENARIO_PACK_URL } from "../config";
+import { loadDefaultScriptEditorTemplateProject } from "../application/default-template-project-loader";
 import type { ScriptEditorProjectDefinition } from "../domain/script-editor-project";
 import {
   createTextImportFilesFromRecord,
@@ -259,11 +258,7 @@ export class ScriptEditorWorkflowController {
   async importTemplateProject(): Promise<void> {
     try {
       this.environment.setProjectSource("imported");
-      this.environment.commitProject(
-        await loadScriptEditorProjectFromScenarioPackUrl(
-          DEFAULT_SCRIPT_EDITOR_TEMPLATE_SCENARIO_PACK_URL
-        )
-      );
+      this.environment.commitProject(await loadDefaultScriptEditorTemplateProject());
       this.environment.resetRecordListPages();
       this.environment.resetRecordSearch();
       this.environment.setSelection({